Improving slot events management
Because events cannot be saved and restore, but we need to keep them, the events now need to be defined somewhere else.

I added an open abtract method `open fun onSlotInitiated( slot: Slot, ... )` in the Node.class . This method is called after a slot is added to the ViewGroup. This allow to give more power to controls slots events, and allow save and restore states.